requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
pymysql==1.1.1
aiohttp==3.10.5
openai==1.51.0

alerts-in-ua==0.3.2
